NF: move Decks in a package
