+++ AnkiDroid/src/main/java/com/ichi2/libanki/Storage.java
@@ -20,6 +20,7 @@ import android.content.ContentValues;
 import android.content.Context;
 
 import com.ichi2.anki.exception.ConfirmModSchemaException;
+import com.ichi2.libanki.decks.Decks;
 
 import com.ichi2.utils.JSONArray;
 import com.ichi2.utils.JSONException;
