+++ AnkiDroid/src/main/java/com/ichi2/libanki/stats/AdvancedStatistics.java
@@ -28,7 +28,7 @@ import com.ichi2.anki.R;
 import com.ichi2.anki.stats.StatsMetaInfo;
 import com.ichi2.libanki.Collection;
 import com.ichi2.libanki.Consts;
-import com.ichi2.libanki.Decks;
+import com.ichi2.libanki.decks.Decks;
 import com.ichi2.libanki.stats.Stats;
 
 import com.ichi2.utils.JSONObject;
